backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
dnspython==2.4.2
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0
email-validator==2.1.0